fixed a problem with validator added for wrong constraint HV000243: Constraint com.infobip.validation.multiple.FirstCustomValidation references constraint validator type com.infobip.validation.multiple.SecondCustomValidationLongValidator, but this validator is defined for constraint type com.infobip.validation.multiple.SecondCustomValidation.
