+++ infobip-bean-validation-spring-boot-starter/src/test/java/com/infobip/validation/multiple/FirstCustomValidation.java
@@ -12,7 +12,7 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 @Target({FIELD})
 @Retention(RUNTIME)
 @Documented
-@interface CustomValidation {
+@interface FirstCustomValidation {
 
     String message() default "";
