+++ infobip-bean-validation-spring-boot-starter/src/test/java/com/infobip/validation/multiple/IntegerWrapper.java
@@ -5,6 +5,6 @@ import lombok.Value;
 @Value
 class IntegerWrapper {
 
-    @CustomValidation
+    @FirstCustomValidation
     private final Integer value;
 }
