+++ infobip-bean-validation-spring-boot-starter/src/test/java/com/infobip/validation/multiple/LongWrapper.java
@@ -5,6 +5,6 @@ import lombok.Value;
 @Value
 class LongWrapper {
 
-    @CustomValidation
+    @FirstCustomValidation
     private final Long value;
 }
