+++ infobip-bean-validation-spring-boot-starter/src/test/java/com/infobip/validation/multiple/SecondCustomValidation.java
@@ -1,0 +1,21 @@
+package com.infobip.validation.multiple;
+
+import javax.validation.Constraint;
+import javax.validation.Payload;
+import java.lang.annotation.*;
+
+import static java.lang.annotation.ElementType.FIELD;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+@Constraint(validatedBy = {})
+@Target({FIELD})
+@Retention(RUNTIME)
+@Documented
+@interface SecondCustomValidation {
+
+    String message() default "";
+
+    Class<?>[] groups() default {};
+
+    Class<? extends Payload>[] payload() default {};
+}
